Move PeerNodeModel to Stratis.Bitcoin.Connection
